[Chore] domain 패지키로 이동
