--- src/main/java/bowling/Frame.java
@@ -1,19 +1,0 @@
-package bowling;
-
-
-import bowling.annotations.ForUI;
-
-public interface Frame {
-    Frame next();
-
-    void bowl(int knockedOutCount);
-
-    boolean isEnd();
-
-    boolean isBeforeFinalFrame();
-
-    boolean isFinalFrame();
-
-    @ForUI
-    KnockedPinCounts getKnockedPinCounts();
-}
