+++ src/main/java/bowling/domain/NormalFrame.java
@@ -2,9 +2,8 @@ package bowling.domain;
 
 public class NormalFrame extends AbstractFrame {
     private final FrameRoundNumber roundNumber;
-    private Frame next;
 
-    public NormalFrame(int number) {
+    private NormalFrame(int number) {
         this(new FrameRoundNumber(number));
     }
 
@@ -17,33 +16,26 @@ public class NormalFrame extends AbstractFrame {
         return new NormalFrame(1);
     }
 
+    public static Frame ofNext(Frame frame) {
+        return frame.next();
+    }
+
     public static Frame ofFinal() {
         return new FinalFrame();
     }
 
     @Override
-    public Frame addNextFrame() {
-        if (isNinthFrame()) {
-            next = NormalFrame.ofFinal();
-            return next;
-        }
-
-        next = new NormalFrame(roundNumber.next());
-        return next;
+    public Frame next() {
+        return new NormalFrame(roundNumber.next());
     }
 
     @Override
-    public boolean isNinthFrame() {
-        return roundNumber.equals(FrameRoundNumber.NINTH_FRAME_NUMBER);
+    public boolean isBeforeFinalFrame() {
+        return roundNumber.equals(FrameRoundNumber.BEFORE_FINAL_FRAME_NUMBER);
     }
 
     @Override
     public boolean isFinalFrame() {
         return false;
-    }
-
-    @Override
-    public Frame next() {
-        return next;
     }
 }
