+++ src/test/java/bowling/FinalFrameTest.java
@@ -1,5 +1,6 @@
 package bowling;
 
+import bowling.domain.FinalFrame;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
