+++ src/test/java/bowling/FrameRoundNumberTest.java
@@ -1,5 +1,6 @@
 package bowling;
 
+import bowling.domain.FrameRoundNumber;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
