+++ src/test/java/bowling/KnockedPinCountTest.java
@@ -1,5 +1,6 @@
 package bowling;
 
+import bowling.domain.KnockedPinCount;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
