+++ src/test/java/bowling/PlayerTest.java
@@ -1,5 +1,6 @@
 package bowling;
 
+import bowling.domain.Player;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
